[FEATURE][TAS-148] Refactor dashboard views: split in different files
