+++ dashboard/mixins.py
@@ -1,0 +1,7 @@
+from django.contrib.auth.mixins import UserPassesTestMixin
+
+
+class SuperuserRequiredMixin(UserPassesTestMixin):
+
+    def test_func(self):
+        return self.request.user.is_superuser
