+++ dashboard/templates/dashboard/layout/base.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-{% load i18n django_vite %}
+{% load i18n static django_vite %}
 <html lang="en" dir="ltr">
   <head>
     <meta charset="UTF-8">
@@ -12,8 +12,7 @@
     <meta name="mobile-web-app-capable" content="yes">
     <meta name="HandheldFriendly" content="True">
     <meta name="MobileOptimized" content="320">
-    <link rel="icon" href="./favicon.ico" type="image/x-icon"/>
-    <link rel="shortcut icon" type="image/x-icon" href="./favicon.ico" />
+    <link rel="icon" href="{% static 'images/favicon.ico' %}" type="image/x-icon"/>
 
     <title>Dashboard Desparchado</title>
