+++ dashboard/urls.py
@@ -22,7 +22,7 @@ urlpatterns = [
         dashboard_views.BlaaEventsListView.as_view(),
         name='blaa_events_list',
     ),
-    path('events/add/', dashboard_views.EventCreateView.as_view(), name='event_add'),
+    path('events/add/', dashboard_views.BlaaEventCreateView.as_view(), name='event_add'),
     # FILBo
     path(
         'filbo/events/create/',
