+++ dashboard/views/__init__.py
@@ -1,0 +1,5 @@
+from .blaa_events import BlaaEventCreateView, BlaaEventsListView
+from .home import HomeView
+from .places import PlacesListView
+from .social import SocialPostsListView, social_events_source
+from .filbo_events import FilboEventFormView
